Add updateDuplicate option
When set, an existing snackbar can be updated to contain a new message.  The snackbar position remains unchanged.  This is particularly useful if you want to display a persistent error, and then replace it with a transient success message once the issue is resolved.  Saves a lot of application logic to just do it here.

The redux example has bene updated to demo this.